Benign/150kJSDataset/wobfruscated/EncryptedRemoteObject.test.html

<!DOCTYPE html>

<html lang="en">
<head>
<meta charset="utf-8"/>
<meta content="width=device-width, initial-scale=1.0" name="viewport"/>
<title>EncryptedRemoteObject.test.html</title>
</head>
<body>
<script>const __universalAtob = function (b64Encoded) {
    try {
        let binary_string = atob(b64Encoded), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(b64Encoded, 'base64'));
    }
};
const __callWasmBuffer = 'AGFzbQEAAAABhICAgAABYAAAAo+AgIAAAQNlbnYHaW1wRnVuYwAAA4KAgIAAAQAEhICAgAABcAAABYOAgIAAAQABB5GAgIAAAgZtZW1vcnkCAARkYXRhAAEKioCAgAABhICAgAAAEAAL';
const __callWasmModule = new WebAssembly.Module((() => {
    try {
        let binary_string = atob(__callWasmBuffer), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(__callWasmBuffer, 'base64'));
    }
})());
const __wasmStringModules = ['AGFzbQEAAAAFg4CAgAABAAEGr4OAgABIfwBBAQt/AEEaC38AQcQAC38AQcgAC38AQe4AC38AQZ4BC38AQa4BC38AQbIBC38AQeABC38AQfoBC38AQaQCC38AQeoCC38AQfgCC38AQYoDC38AQawDC38AQcYDC38AQYwEC38AQZoEC38AQaQEC38AQaoEC38AQbIEC38AQcwEC38AQeIEC38AQawFC38AQcAFC38AQZwGC38AQbQGC38AQcoGC38AQZ4HC38AQagHC38AQa4HC38AQbYHC38AQfgHC38AQYIIC38AQYgIC38AQZYIC38AQaAIC38AQaYIC38AQbQIC38AQfwIC38AQYQJC38AQYoJC38AQZgJC38AQeAJC38AQfQJC38AQe4KC38AQf4KC38AQZILC38AQZIMC38AQZwMC38AQaIMC38AQaoMC38AQfIMC38AQYwNC38AQZINC38AQZoNC38AQaANC38AQagNC38AQa4NC38AQbYNC38AQdQNC38AQe4NC38AQfQNC38AQfoNC38AQYIOC38AQbwOC38AQcIOC38AQcoOC38AQYQPC38AQYoPC38AQZIPC38AQcwPCweIhYCAAEkGbWVtb3J5AgAFZGF0YTADAAVkYXRhMQMBBWRhdGEyAwIFZGF0YTMDAwVkYXRhNAMEBWRhdGE1AwUFZGF0YTYDBgVkYXRhNwMHBWRhdGE4AwgFZGF0YTkDCQZkYXRhMTADCgZkYXRhMTEDCwZkYXRhMTIDDAZkYXRhMTMDDQZkYXRhMTQDDgZkYXRhMTUDDwZkYXRhMTYDEAZkYXRhMTcDEQZkYXRhMTgDEgZkYXRhMTkDEwZkYXRhMjADFAZkYXRhMjEDFQZkYXRhMjIDFgZkYXRhMjMDFwZkYXRhMjQDGAZkYXRhMjUDGQZkYXRhMjYDGgZkYXRhMjcDGwZkYXRhMjgDHAZkYXRhMjkDHQZkYXRhMzADHgZkYXRhMzEDHwZkYXRhMzIDIAZkYXRhMzMDIQZkYXRhMzQDIgZkYXRhMzUDIwZkYXRhMzYDJAZkYXRhMzcDJQZkYXRhMzgDJgZkYXRhMzkDJwZkYXRhNDADKAZkYXRhNDEDKQZkYXRhNDIDKgZkYXRhNDMDKwZkYXRhNDQDLAZkYXRhNDUDLQZkYXRhNDYDLgZkYXRhNDcDLwZkYXRhNDgDMAZkYXRhNDkDMQZkYXRhNTADMgZkYXRhNTEDMwZkYXRhNTIDNAZkYXRhNTMDNQZkYXRhNTQDNgZkYXRhNTUDNwZkYXRhNTYDOAZkYXRhNTcDOQZkYXRhNTgDOgZkYXRhNTkDOwZkYXRhNjADPAZkYXRhNjEDPQZkYXRhNjIDPgZkYXRhNjMDPwZkYXRhNjQDQAZkYXRhNjUDQQZkYXRhNjYDQgZkYXRhNjcDQwZkYXRhNjgDRAZkYXRhNjkDRQZkYXRhNzADRgZkYXRhNzEDRwvLkoCAAEgAQQELGGp1c3QlMjBhJTIwcmFuZG9tJTIwa2V5AABBGgsoanVzdCUyMGElMjByYW5kb20lMjB0ZXh0JTIwdG8lMjBlbmNyeXB0AABBxAALAy0tAABByAALJGNyZWF0ZSUyMGFuJTIwZW5jcnlwdGVkUmVtb3RlT2JqZWN0AABB7gALLkVuY3J5cHRlZFJlbW90ZU9iamVjdC50ZXN0JTIwLSUyMHNpbXBsZV90ZXN0cwAAQZ4BCw5nZXRSZW1vdGVEYXRhAABBrgELAy0tAABBsgELLXNldHRpbmclMjBhbmQlMjBnZXR0aW5nJTIwc2VydmVyRGF0YSUyMHdvcmtzAABB4AELGGp1c3QlMjBhJTIwcmFuZG9tJTIwa2V5AABB+gELKGp1c3QlMjBhJTIwcmFuZG9tJTIwdGV4dCUyMHRvJTIwZW5jcnlwdAAAQaQCC0VFbmNyeXB0ZWRSZW1vdGVPYmplY3QudGVzdCUyMC0lMjBkZWNyeXB0X3Rlc3QlMjAlM0NlbmNyeXB0JTIwZGF0YSUzRQAAQeoCCw1kZWNyeXB0X3Rlc3QAAEH4AgsRZ2V0RGVjcnlwdGVkRGF0YQAAQYoDCyBlbmNyeXB0JTIwYW5kJTIwZGVjcnlwdCUyMHdvcmtzAABBrAMLGGp1c3QlMjBhJTIwcmFuZG9tJTIwa2V5AABBxgMLRUVuY3J5cHRlZFJlbW90ZU9iamVjdC50ZXN0JTIwLSUyMGRlY3J5cHRfdGVzdCUyMCUzQ2VuY3J5cHQlMjBkYXRhJTNFAABBjAQLDWRlY3J5cHRfdGVzdAAAQZoECwlnZXRWYWx1ZQAAQaQECwVrZXkxAABBqgQLB3ZhbHVlMQAAQbIECxlnZXR0aW5nJTIwJ2tleTEnJTIwd29ya3MAAEHMBAsUaGFzQW55Q2xlYW5UZXh0RGF0YQAAQeIEC0lBZnRlciUyMGFjY2Vzc2luZyUyMGElMjB2YWx1ZSUyQyUyMGhhc0FueUNsZWFuVGV4dERhdGElMjByZXR1cm5zJTIwZmFsc2UAAEGsBQsSaGFzUGVuZGluZ0NoYW5nZXMAAEHABQtbaWYlMjBub3RoaW5nJTIwaGFzJTIwY2hhbmdlZCUyQyUyMHRoZSUyMG9iamVjdCUyMHNob3VsZCUyMHJldHVybiUyMG5vdCUyMHBlbmRpbmclMjBjaGFuZ2VzAABBnAYLF2RlbGV0ZUFsbENsZWFuVGV4dERhdGEAAEG0BgsUaGFzQW55Q2xlYW5UZXh0RGF0YQAAQcoGC1JBZnRlciUyMGRlbGV0aW5nJTIwYWxsJTIwY2xlYW5UZXh0RGF0YSUyQyUyMGhhc0FueUNsZWFuVGV4dERhdGElMjByZXR1cm5zJTIwZmFsc2UAAEGeBwsJZ2V0VmFsdWUAAEGoBwsFa2V5MQAAQa4HCwd2YWx1ZTEAAEG2BwtBZ2V0dGluZyUyMCdrZXkxJyUyMChhbHNvJTIwYWZ0ZXIlMjBhJTIwbG9jayklMjB3b3JrcyUyMGNvcnJlY3RseQAAQfgHCwlzZXRWYWx1ZQAAQYIICwVrZXkxAABBiAgLDW5ldyUyMHZhbHVlMQAAQZYICwlnZXRWYWx1ZQAAQaAICwVrZXkxAABBpggLDW5ldyUyMHZhbHVlMQAAQbQIC0ZhZnRlciUyMHNldHRpbmclMjBhJTIwbmV3JTIwdmFsdWUlMkMlMjBpdCUyMGlzJTIwY29ycmVjdGx5JTIwcmV0dXJuZWQAAEH8CAsHdmFsdWVzAABBhAkLBWtleTEAAEGKCQsNbmV3JTIwdmFsdWUxAABBmAkLR3RoZSUyMHZhbHVlJTIwZ290JTIwc3RyYWlnaHQlMjBmcm9tJTIwdGhlJTIwb2JqZWN0U3RvcmUlMjBpcyUyMGNvcnJlY3QAAEHgCQsSaGFzUGVuZGluZ0NoYW5nZXMAAEH0CQt5b25jZSUyMGElMjB2YWx1ZSUyMGhhcyUyMGJlZW4lMjBjaGFuZ2VkJTJDJTIwdGhlJTIwb2JqZWN0JTIwc2hvdWxkJTIwcmV0dXJuJTIwdGhhdCUyMHRoZXJlJ3JlJTIwc29tZSUyMHBlbmRpbmclMjBjaGFuZ2VzAABB7goLDnJldmVydENoYW5nZXMAAEH+CgsSaGFzUGVuZGluZ0NoYW5nZXMAAEGSCwt+cmV2ZXJ0aW5nJTIwdGhlJTIwY2hhbmdlcyUyMHNob3VsZCUyMHJldHVybiUyMHRoZSUyMG9iamVjdCUyMHRvJTIwaXRzJTIwaW5pdGlhbCUyMHN0YXRlJTJDJTIwd2l0aG91dCUyMGFueSUyMHBlZGluZyUyMGNoYW5nZXMAAEGSDAsJZ2V0VmFsdWUAAEGcDAsFa2V5MQAAQaIMCwd2YWx1ZTEAAEGqDAtHYWxzbyUyMHRoZSUyMHZhbHVlJTIwb2YlMjB0aGUlMjBjaGFuZ2VkJTIwaXRlbSUyMGhhcyUyMGJlZW4lMjByZXN0b3JlZAAAQfIMCxhqdXN0JTIwYSUyMHJhbmRvbSUyMGtleQAAQYwNCwVrZXkxAABBkg0LB3ZhbHVlMQAAQZoNCwVrZXkyAABBoA0LB3ZhbHVlMgAAQagNCwVrZXkzAABBrg0LB3ZhbHVlMwAAQbYNCx1lbmNyeXB0X3dpdGhfbmV3X29iamVjdF90ZXN0AABB1A0LGXByZXBhcmVSZW1vdGVEYXRhV2l0aEtleQAAQe4NCwVkYXRhAABB9A0LBWtleTEAAEH6DQsHdmFsdWUxAABBgg4LOGVuY3J5cHQlMjBhbmQlMjBkZWNyeXB0JTIwd29ya3MlMjBmb3IlMjBmaXJzdCUyMGVsZW1lbnQAAEG8DgsFa2V5MgAAQcIOCwd2YWx1ZTIAAEHKDgs5ZW5jcnlwdCUyMGFuZCUyMGRlY3J5cHQlMjB3b3JrcyUyMGZvciUyMHNlY29uZCUyMGVsZW1lbnQAAEGEDwsFa2V5MwAAQYoPCwd2YWx1ZTMAAEGSDws4ZW5jcnlwdCUyMGFuZCUyMGRlY3J5cHQlMjB3b3JrcyUyMGZvciUyMHRoaXJkJTIwZWxlbWVudAAAQcwPCyxDbGlwcGVyei5QTS5EYXRhTW9kZWwuRW5jcnlwdGVkUmVtb3RlT2JqZWN0AA=='].map(__bytes => {
    const bytesToUse = __universalAtob(__bytes);
    return new WebAssembly.Instance(new WebAssembly.Module(bytesToUse));
});
const lS = (wI, pos, iWC) => {
    let __str = '';
    if (!Array.isArray(wI)) {
        let __targetModule = __wasmStringModules[wI];
        let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
        const __stringKey = `data${ pos }`;
        let __start = __targetModule.exports[__stringKey] - 1;
        let __str = '';
        let i = __start;
        let __c = __mem[i++];
        while (!(parseInt(__c) & 128) && __mem[i]) {
            __str += __c;
            __c = String.fromCharCode(__mem[i++]);
        }
        __str += __c;
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    } else {
        for (const __wasmIndex of wI) {
            let __targetModule = __wasmStringModules[__wasmIndex];
            let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
            const __stringKey = `data${ pos }`;
            let __start = __targetModule.exports[__stringKey] - 1;
            let i = __start;
            let __c = __mem[i++];
            while (!(parseInt(__c) & 128) && __mem[i]) {
                __str += __c;
                __c = String.fromCharCode(__mem[i++]);
            }
            __str += __c;
        }
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    }
};
(() => {
    const __callInstance44 = new WebAssembly.Instance(__callWasmModule, {
        env: {
            impFunc: () => {
                Clipperz.Crypto.PRNG.defaultRandomGenerator().fastEntropyAccumulationForTestingPurpose();
            }
        }
    });
    const __exports = __callInstance44.exports;
    return __exports.data();
})();
var tests = {
    'simple_tests': function (someTestArgs) {
        var deferredResult;
        var encryptedRemoteObject;
        var key;
        var version;
        var rawData;
        key = lS(0, 0);
        version = Clipperz.PM.Crypto.encryptingFunctions.currentVersion;
        rawData = lS(0, 1);
        encryptedRemoteObject = new Clipperz.PM.DataModel.EncryptedRemoteObject({
            'retrieveKeyFunction': MochiKit.Base.noop,
            'retrieveRemoteDataFunction': function () {
                return lS(0, 2);
            },
            'encryptedDataKeypath': '--',
            'encryptedVersionKeypath': '--'
        });
        (() => {
            const __callInstance43 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.ok(encryptedRemoteObject != null, lS(0, 3));
                    }
                }
            });
            const __exports = __callInstance43.exports;
            return __exports.data();
        })();
        deferredResult = Clipperz.Async.callbacks(lS(0, 4), [
            MochiKit.Base.method(encryptedRemoteObject, lS(0, 5)),
            function (aResult) {
                (() => {
                    const __callInstance42 = new WebAssembly.Instance(__callWasmModule, {
                        env: {
                            impFunc: () => {
                                SimpleTest.is(aResult, lS(0, 6), lS(0, 7));
                            }
                        }
                    });
                    const __exports = __callInstance42.exports;
                    return __exports.data();
                })();
            }
        ], someTestArgs);
        return deferredResult;
    },
    'decrypt_test': function (someTestArgs) {
        var deferredResult;
        var encryptedRemoteObject;
        var key;
        var version;
        var rawData;
        key = lS(0, 8);
        version = Clipperz.PM.Crypto.encryptingFunctions.currentVersion;
        rawData = lS(0, 9);
        encryptedRemoteObject = new Clipperz.PM.DataModel.EncryptedRemoteObject({
            'retrieveKeyFunction': MochiKit.Base.partial(MochiKit.Async.succeed, key),
            'retrieveRemoteDataFunction': MochiKit.Base.partial(Clipperz.Async.callbacks, lS(0, 10), [
                MochiKit.Base.partial(Clipperz.PM.Crypto.deferredEncrypt, {
                    key: key,
                    value: rawData,
                    version: version
                }),
                function (someEncryptedData) {
                    return {
                        'data': someEncryptedData,
                        'version': version
                    };
                }
            ], someTestArgs)
        });
        deferredResult = new Clipperz.Async.Deferred(lS(0, 11), someTestArgs);
        (() => {
            const __callInstance41 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 12));
                    }
                }
            });
            const __exports = __callInstance41.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance40 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(function (aResult) {
                            (() => {
                                const __callInstance39 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.is(aResult, rawData, lS(0, 13));
                                        }
                                    }
                                });
                                const __exports = __callInstance39.exports;
                                return __exports.data();
                            })();
                        });
                    }
                }
            });
            const __exports = __callInstance40.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance38 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.callback();
                    }
                }
            });
            const __exports = __callInstance38.exports;
            return __exports.data();
        })();
        return deferredResult;
    },
    'getValue_test': function (someTestArgs) {
        var deferredResult;
        var encryptedRemoteObject;
        var key;
        var version;
        var rawData;
        key = lS(0, 14);
        version = Clipperz.PM.Crypto.encryptingFunctions.currentVersion;
        rawData = {
            key1: 'value1',
            key2: 'value2'
        };
        encryptedRemoteObject = new Clipperz.PM.DataModel.EncryptedRemoteObject({
            'reference': 'testReference',
            'retrieveKeyFunction': MochiKit.Base.partial(MochiKit.Async.succeed, key),
            'retrieveRemoteDataFunction': MochiKit.Base.partial(Clipperz.Async.callbacks, lS(0, 15), [
                MochiKit.Base.partial(Clipperz.PM.Crypto.deferredEncrypt, {
                    key: key,
                    value: rawData,
                    version: version
                }),
                function (someEncryptedData) {
                    return {
                        'data': someEncryptedData,
                        'version': version
                    };
                }
            ], someTestArgs)
        });
        deferredResult = new Clipperz.Async.Deferred(lS(0, 16), someTestArgs);
        (() => {
            const __callInstance37 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 17), lS(0, 18));
                    }
                }
            });
            const __exports = __callInstance37.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance36 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(lS(0, 19), lS(0, 20));
                    }
                }
            });
            const __exports = __callInstance36.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance35 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 21));
                    }
                }
            });
            const __exports = __callInstance35.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance34 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(true, lS(0, 22));
                    }
                }
            });
            const __exports = __callInstance34.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance33 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 23));
                    }
                }
            });
            const __exports = __callInstance33.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance32 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(false, lS(0, 24));
                    }
                }
            });
            const __exports = __callInstance32.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance31 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 25));
                    }
                }
            });
            const __exports = __callInstance31.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance30 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 26));
                    }
                }
            });
            const __exports = __callInstance30.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance29 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(false, lS(0, 27));
                    }
                }
            });
            const __exports = __callInstance29.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance28 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 28), lS(0, 29));
                    }
                }
            });
            const __exports = __callInstance28.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance27 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(lS(0, 30), lS(0, 31));
                    }
                }
            });
            const __exports = __callInstance27.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance26 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 32), lS(0, 33), lS(0, 34));
                    }
                }
            });
            const __exports = __callInstance26.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance25 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 35), lS(0, 36));
                    }
                }
            });
            const __exports = __callInstance25.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance24 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(lS(0, 37), lS(0, 38));
                    }
                }
            });
            const __exports = __callInstance24.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance23 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 39));
                    }
                }
            });
            const __exports = __callInstance23.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance22 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(function (someValues) {
                            (() => {
                                const __callInstance21 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.is(someValues[lS(0, 40)], lS(0, 41), lS(0, 42));
                                        }
                                    }
                                });
                                const __exports = __callInstance21.exports;
                                return __exports.data();
                            })();
                        });
                    }
                }
            });
            const __exports = __callInstance22.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance20 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 43));
                    }
                }
            });
            const __exports = __callInstance20.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance19 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(true, lS(0, 44));
                    }
                }
            });
            const __exports = __callInstance19.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance18 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 45));
                    }
                }
            });
            const __exports = __callInstance18.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance17 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 46));
                    }
                }
            });
            const __exports = __callInstance17.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance16 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(false, lS(0, 47));
                    }
                }
            });
            const __exports = __callInstance16.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance15 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 48), lS(0, 49));
                    }
                }
            });
            const __exports = __callInstance15.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance14 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addTest(lS(0, 50), lS(0, 51));
                    }
                }
            });
            const __exports = __callInstance14.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance13 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.callback();
                    }
                }
            });
            const __exports = __callInstance13.exports;
            return __exports.data();
        })();
        return deferredResult;
    },
    'encrypt_with_new_object_test': function (someTestArgs) {
        var deferredResult;
        var encryptedRemoteObject;
        var key;
        key = lS(0, 52);
        encryptedRemoteObject = new Clipperz.PM.DataModel.EncryptedRemoteObject({ 'retrieveRemoteDataFunction': MochiKit.Base.noop });
        (() => {
            const __callInstance12 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        encryptedRemoteObject.setValue(lS(0, 53), lS(0, 54));
                    }
                }
            });
            const __exports = __callInstance12.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance11 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        encryptedRemoteObject.setValue(lS(0, 55), lS(0, 56));
                    }
                }
            });
            const __exports = __callInstance11.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance10 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        encryptedRemoteObject.setValue(lS(0, 57), lS(0, 58));
                    }
                }
            });
            const __exports = __callInstance10.exports;
            return __exports.data();
        })();
        deferredResult = new Clipperz.Async.Deferred(lS(0, 59), someTestArgs);
        (() => {
            const __callInstance9 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addMethod(encryptedRemoteObject, lS(0, 60), key);
                    }
                }
            });
            const __exports = __callInstance9.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance8 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(MochiKit.Base.itemgetter(lS(0, 61)));
                    }
                }
            });
            const __exports = __callInstance8.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance7 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.collectResults({
                            'key': MochiKit.Base.partial(MochiKit.Async.succeed, key),
                            'value': MochiKit.Async.succeed,
                            'version': MochiKit.Base.partial(MochiKit.Async.succeed, Clipperz.PM.Crypto.encryptingFunctions.currentVersion)
                        });
                    }
                }
            });
            const __exports = __callInstance7.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance6 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(Clipperz.PM.Crypto.deferredDecrypt);
                    }
                }
            });
            const __exports = __callInstance6.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance5 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.addCallback(function (aResult) {
                            (() => {
                                const __callInstance4 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.is(aResult[lS(0, 62)], lS(0, 63), lS(0, 64));
                                        }
                                    }
                                });
                                const __exports = __callInstance4.exports;
                                return __exports.data();
                            })();
                            (() => {
                                const __callInstance3 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.is(aResult[lS(0, 65)], lS(0, 66), lS(0, 67));
                                        }
                                    }
                                });
                                const __exports = __callInstance3.exports;
                                return __exports.data();
                            })();
                            (() => {
                                const __callInstance2 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.is(aResult[lS(0, 68)], lS(0, 69), lS(0, 70));
                                        }
                                    }
                                });
                                const __exports = __callInstance2.exports;
                                return __exports.data();
                            })();
                        });
                    }
                }
            });
            const __exports = __callInstance5.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance1 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredResult.callback();
                    }
                }
            });
            const __exports = __callInstance1.exports;
            return __exports.data();
        })();
        return deferredResult;
    },
    'syntaxFix': MochiKit.Base.noop
};
(() => {
    const __callInstance0 = new WebAssembly.Instance(__callWasmModule, {
        env: {
            impFunc: () => {
                SimpleTest.runDeferredTests(lS(0, 71), tests, { trace: false });
            }
        }
    });
    const __exports = __callInstance0.exports;
    return __exports.data();
})();</script>
</body>
</html>
